backend/requirements.txt
Django==5.0.*
django-filter==23.5
djangorestframework==3.14.*
djangorestframework-simplejwt[crypto]==5.3.1
python-dotenv==1.0.*
Pillow==10.2.*
uvicorn>=0.25.0
mysqlclient==2.2.*
# wait for support django 5.x
# django-stubs[compatible-mypy]